Allow Doom to build and run on FreeBSD
The existing i_video.c module doesn't work on modern systems anymore
because it insists on using an 8-bit PseudoColor X11 visual. Most
systems no longer support that. In order to address this, the code has
been updated to support using a 24-bit TrueColor X11 visual instead.

Also, a Makefile has been added to build Doom on FreeBSD. The default
is to build without sound support, but not music, since that requires
the user to install the WildMidi library and set up their wildmidi.cfg
file and Gravis Ultrasound patches.

All of the code for the badge is built on FreeBSD except for the
i_video.c and i_sound.c modules (and a tiny number of things covered
by the BADGEDOOM macro). This means the BSD and badge builds of Doom
should be bug-for-bug compatible.