Fix another WildMidi bug
The WM_GetOutput_Gauss() and WM_GetOutput_Linear() routines are
designed to implement looping by checking for an end-of-track
message in the event list. The assumption is that this will
indicate the end of the tune and if we see it, and the WM_MO_LOOP
flag is set, we should loop back to the start of the tune.

But this assumes that there can only ever be one end-of-track
message in a MIDI file, and that assumption is apparently wrong:
it discounts the possibility of a single MIDI tune file containing
multiple tracks. In some cases, there may be a short track at the
start of of the tune with some non-note events that terminates early.
Currently, the GetOutput routines treat these early end-of-track
events as end of song markers, and if the WM_MO_LOOP flag is set,
the wull use _WM_ResetToStart() to rewind back to the start of the
tune and process them again. But this causes them to get stuck
looping over these non-note regions forever, and they never return.

To address this, the logic has been made more specific so that the
end-of-track event is only honored if it is also the last valid
event in the event list for this tune (i.e. the event immediately
following is the list terminator, with an evtype of ev_null). This
more accurately reflects the actual end of song.

I'm not sure if this is something that can happen with actual MUS
tunes; so far I've only confirmed it happening with newer Doom PWADs
that use standard MIDI tunes instead. These would not normally be
compatible with vanilla Doom, but happen to work here since
WildMidi supports both formats automagically.