Fix Doom's aspect ratio
You probably remember Doom as having VGA 640x480 graphics, but
it turns out that's not true. In reality, it used something
called VGA mode Y, which actually used a 320x200 resolution.
This is why Doom's rendering code is configured to produce
graphics frames with this resolution instead of, say 320x240,
which seems closer to actual VGA resolution.

It turns out that VGA display controllers would stretch the
pixels when operating in 320x200 mode in order to fill the
entire screen and still present a 4.3 aspect ratio, even
though 320x200 seems like it would yield an aspect ratio of
4:2.5 instead. (And it would, if the pixels were square.)

Furthermore, the game sprites like monsters and Doomguy's face
in the HUD were intentionally drawn so that they would appear
to be correcly proportioned once they were stretched.

One assumes this done for performance reasons, since rendering
320x200 pixels takes a little less time than drawing 320x240
(and certainly less than 640x480).

The problem is that the LCD display on the STM32F746 reference
board and the displays we need to use for the production badge
design have square pixels. We can display in 320x240 resolution
which does provide a 4:3 aspect ratio, and we can directly draw
the 320x200 frames from the Doom rendering like that without any
problem, except the frames are 40 pixels too short and appear
squashed.

Unfortunately there doesn't seem to be any hardware gimmick
we can use to mitigate this, so we have to settle for a
software workaround. The badge-specific video module will now
scale the display frames from 4:2.5 to 4:3 aspect ratio by
duplicating every 5th scanline. Fortunartely the CPU seems
fast enough to handle this without hurting game performance.

The same hack is also implemented for the UNIX i_video.c
module, just for kicks.