Fix some NES emulator issues
When calling vid_blit(), we need to make sure to flush
the original bitmap data from the cache before processing
it with the DMA2D engine. Failing to do this sometimes
results in misdrawn graphics (example: in the intro video
in Battletoads, when the Black Queen first appears and
the screen goes "static-y" as she interrupts the playback,
there is glitching near the bottom of the screen).

Also, we need to increase the audio sample buffer count
a little to ensure audio playback for a given video frame
always takes slightly longer than the time needed to render
the video.