Re-skin app-email.c again
The "Bitcoin Miner" app is now an X app, in (dis)honor of you
know who. As usual all it does is Rickroll you.

Also, the launcher and ui-list modules now use Doom sound
effects instead of the old submarine effects from the DC27
design. I'm not sure if they will stay this way for good,
but I was getting bored of the old sounds.