Fix problem with defaults handling on Windows
The UNIX platforms we build on use the LP64 model for 64-bit build,
meaning that longs and pointers are 64 bits in size and ints remain
32-bits in size. But Windows/MinGW is different: longs are still 32
bits in size. This breaks the assumptions in the defaults handling
code where we try to squeeze a 64-bit pointer in to the defaultvalue
field, which is declared as a long. On LP64 platforms this works,
but not on Windows. We have to explicitly use a long long instead.