Create separate Makefile for MacOS build
We need to make the section(".ram7") attributes dissapear for MacOS,
as it uses the mach-o binary format and the section specifier is not
correct for that. We don't really need the section attribute for
any build besides the badge anyway, so we can force them off. Luckily,
specifying an unsupported attribute only triggers a compiler warning
rather than a hard error, so we can use that approach from the
Makefile rather than editing all of the source.